Create caesar_Cipher.py
The Caesar cypher is one of the most basic and well-known encryption methods. It was called after Julius Caesar, who allegedly used it to communicate with his army's officials. 
The cypher operates by relocating each letter in the plaintext (the original message) down the alphabet by a set number of places. For example, with a key of 3, 'A' would be encoded as 'D', 'B' as 'E', and so on. This shifting goes all the way around the alphabet, thus 'Z' becomes 'C' with a key of 3.
The procedure is reversed to decode the message by moving each letter back by the same amount of places. The encryption strength is determined by the key value. A key of 1 is, for example, a very poor encryption since it is easily guessable, yet a key of 25 is comparable to a key of 1 because it just loops around the alphabet.
Although the Caesar cypher is simple to learn and use, it is not considered safe by contemporary cryptographic standards. It is easily broken using brute force methods, especially with computers readily available. Nonetheless, the Caesar cypher is a fundamental principle for more sophisticated encryption schemes.
The Caesar cypher algorithm is used in this code to encrypt and decode communications. A secret message, an encryption key, and the chosen mode (encrypt/decrypt) must be entered by the user. Depending on the mode and key that have been chosen, the code then goes through the message character by character, encrypting or decrypting symbols. The output is ultimately shown after being saved as the translated message.